Update message flow control to be simpler.
Now the message is processed by a handler at the task level which
parses and validate the received JSON message through the Message
Decoder. This decoder was previously named WrapperJSON but now
Message Decoder makes more sense. Finally, the handler calls the
updateRawCommand function to fill the Raw Command object.

Note, since the handler parse the message the JSON is stored at
the decoder.