fix: controldev_websocket would forever reject new connections if a handshake failed ever happened
msg was unnecessarily defined at the class level. This meant that
whatever data that has ever been written on it would always stick
and stay there forever.

In practice, only the handshake failed connection state would cause
this. We would have to kill the whole component whenever this situation
happens.