fix: race condition in start
The synchronization between the server's startup seequence (which
runs in a thread) and startHook was not waiting for the server to
actually listen. This had two effects: the initialization could
(silently) fail, and the task could be started from the unit tests
point of view, but the server was not yet listening.